requirements.txt
certifi==2025.8.3
elastic-transport==9.1.0
elasticsearch==9.1.1
mariadb==1.1.13
packaging==25.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
